Moving all things that are implementation details but could be reused between architectures into `shim`.
