+++ src/arch/shim/intrin/macros.rs
