+++ src/arch/shim/intrin/popcnt.rs
@@ -5,12 +5,16 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use crate::vecs::*;
-
-pub trait Popcnt : Packed {
-    fn count_ones(&self) -> usize;
-    fn count_zeroes(&self) -> usize {
-        (Self::WIDTH * Self::Scalar::SIZE) - self.count_ones()
+macro_rules! impl_popcnt {
+    ($($vec:ty, $fn:ident),*) => {
+        $(
+            impl Popcnt for $vec {
+                #[inline(always)]
+                #[allow(unused_unsafe)]
+                fn count_ones(&self) -> usize {
+                    unsafe { $fn(self.be_u8s()) }
+                }
+            }
+        )*
     }
 }
-
