+++ src/arch/shim/mod.rs
@@ -1,1 +1,2 @@
+#[macro_use] pub mod intrin;
 pub mod vecs;
