+++ src/stride.rs
@@ -13,8 +13,6 @@ use crate::std::iter::{Iterator, ExactSizeIterator, FromIterator};
 
 // For AVX2 gathers
 use crate::std::mem::transmute;
-use crate::vektor::x86_64::*;
-use crate::vektor::x86::*;
 use crate::intrin::transmute::*;
 
 /// A slice-backed iterator which packs every nth element of its constituent
